+++ main/src/test/java/net/oneandone/stool/configuration/ProjectConfigurationTest.java
@@ -33,7 +33,7 @@ public class ProjectConfigurationTest {
 
         world = World.create();
         gson = Session.gson(world);
-        configuration = new StageConfiguration(null,"refresh");
+        configuration = new StageConfiguration(null);
         tmp = world.getTemp().createTempFile();
         configuration.save(gson, tmp);
         StageConfiguration.load(gson, tmp);